fix: Remove address book context menu
